update alerts to sweet alerts in students.

+++ scripts/student.js
@@ -73,7 +73,7 @@ function addStudent() {
   })
   .then((response) => response.text())
   .then(responseText => {
-    alert(responseText);
+    Swal.fire("Success", `${responseText}`, "success");
     displayStudents(); // updates the table
   }).catch (error => {
     alert('console error.');
@@ -122,7 +122,7 @@ function editStudent(button) {
     })
     .then((response) => response.text())
     .then((responseText) => {
-      alert(responseText);
+      Swal.fire("Success", `${responseText}`, "success");
       displayStudents();
     });
 
@@ -131,6 +131,28 @@ function editStudent(button) {
     // Edit mode
     button.textContent = "Save";
     cells[0].focus();
+
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      icon: "info",
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      customClass: {
+        popup: 'colored-toast'
+      },
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      }
+    });
+
+    Toast.fire({
+      icon: 'info',
+      title: 'Edit Mode Activated'
+    });
+
     updatedRow.length = 0; // clear the array.
   }
 }
@@ -140,19 +162,28 @@ function deleteStudent(button) {
   const row = button.closest('tr');
   const cells = row.querySelectorAll('td');
   const studentId = cells[0].innerHTML;
-  
-  fetch(studentEndpoint, {
-    method: 'DELETE',
-    headers: {
-      "Content-type": "application/x-www-form-urlencoded",
-    },
-    body: `student_id=${studentId}`,
-  })
-  .then((response) => response.text())
-  .then((responseText) => {
-    alert(responseText);
-    displayStudents();
-  })
+
+  Swal.fire({
+    title: "Do you want to delete this data?",
+    icon: "question",
+    showCancelButton: true,
+    confirmButtonText: "Delete",
+  }).then((result) => {
+    if (result.isConfirmed) {
+        fetch(studentEndpoint, {
+          method: 'DELETE',
+          headers: {
+            "Content-type": "application/x-www-form-urlencoded",
+          },
+          body: `student_id=${studentId}`,
+        })
+        .then((response) => response.text())
+        .then(() => {
+          Swal.fire("Deleted", "", "success");
+          displayStudents();
+      })
+    }
+  });
 }
 
 function searchStudentId() {
